requirements.txt
fastapi
uvicorn
transformers
torch
langchain
langchain-anthropic
anthropic
python-multipart
pillow